backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.35.2
sentence-transformers==2.2.2
pymilvus==2.3.5
torch==2.1.1
python-dotenv==1.0.0
pydantic==2.5.0
